fix: reuse default layout config on additional layouts
